pyproject.toml
[project]
name = "crypto-rss-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.12",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-openai>=0.3.33",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "supabase>=2.9.1",
]